+++ bpdv3/Assets/Scripts/Coin/CoinProperties.cs
@@ -24,7 +24,7 @@ public class CoinProperties : MonoBehaviour
         if (other.gameObject.CompareTag("Player"))
         {
             // need to address --- anchorScript.anchorTileObject.GetComponent<TileProperties>().OccupiedDecreased();
-            gameUiGo.GetComponent<GameUI>().Scoring(coinValue);
+            // after change to gameui script -----gameUiGo.GetComponent<GameUI>().Scoring(coinValue);
             Spawner.HasCoinSpawned = false;
             gameObject.SetActive(false);            
         }
