+++ bpdv3/Assets/Scripts/UI/BeatBar.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1645dae497e543545a4854a496dc47bd
+guid: 5cbbf088077142441a098042507634c1
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
