+++ bpdv3/Assets/Scripts/UI/Lives.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1264dd32a2b9d7e4d85fd011769e4da4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
