+++ bpdv3/Assets/Scripts/UI/Scoring.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb190fba1a099704b87e0e3299aa4d13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
